Move pkg/config/remote/uptane -> pkg/remoteconfig/uptane
In order to create a backend client you need to be able to create
an uptane client and thus these two packages are tied together and
should either be both exported in the same module or both modules. We
want to move all our code here anyways, so while we figure it out go
ahead and move it.