Fix issue 2:
One possible solution for what you are trying to do is to use bpf_memcpy
to copy a known chunk into the transaction->topic_name buffer. Then use
the computed topic_name_size value to add a NULL byte at the end of the
string. This way you don't have to use the slower helper bpf_probe_read
to read a variable byte chunk, and you don't have to use a for-loop
either.