pkg/trace/event: fix flaky test
This test is inherently probabilistic, since we're measuring statistical
outcomes. However, in this PR, we bump up the sample size to ensure it
should basically never fall out of range. This does not ensure it will
never fail (we can't without redesigning the test), but it will at least
not flake out all the time.

This takes the test from failing roughly 1/30 times to failing < 1/10000
times - I couldn't get a failure by rerunning the test 10000 times.